Increase EFI image size for El Torito boot method
Current build leaves ~800K free of 31M in the FAT filesystem, adding 9M.

Going beyond 65535 sectors of 512-byte is a bit special,
but works for EFI. Image size is reported a zero in boot catalog,
but xorriso does a good job, and writes right value in hybrid-partition.

Tested booting in qemu in both modes (cd-rom and hybrid) and works fine.

Signed-off-by: Gerardo Exequiel Pozzi <vmlinuz386@gmail.com>